encoding/bom-handling.html

<!-- starts with a BOM -->
<!doctype html>
<meta charset="windows-1252">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<title>BOM handling</title>
<div id=log></div>
<div id=content>これは文字実験です。</div>
<script>

test(
  () => {
    assert_equals(
      document.getElementById("content").textContent,
      "\u3053\u308C\u306F\u6587\u5B57\u5B9F\u9A13\u3067\u3059\u3002"
    );
  },
  "The BOM should override any other encoding declarations."
);

test(
  () => {
    assert_equals(document.characterSet.toLowerCase(), "utf-8");
  },
  "document.characterSet should match the BOM"
);

</script>
